requirements.txt
python-dotenv
fastapi[all]
uvicorn
pipecat-ai[daily,elevenlabs,openai,silero,cerebras,azure]
loguru
onnxruntime